First & Second page of app
First page will direct user to the website that our team created. 
When user clicks on "Login" button, it will direct user to second page to login
Main reason for asking users to login is to make sure whoever uses this application is SJSU student or faculty members for campus security.